Move OAuthMockDelegatingHandler to fixtures
