+++ src/DiscogsApiClient.Tests/Authentication/DiscogsAuthenticationServiceTests.cs
@@ -1,6 +1,5 @@
 using DiscogsApiClient.Authentication.OAuth;
 using DiscogsApiClient.Authentication.PersonalAccessToken;
-using DiscogsApiClient.Tests.MockMiddleware;
 
 namespace DiscogsApiClient.Tests.Authentication;
